move handler creation to pkg node, remove deprecated function
