+++ rpc/gzip.go
@@ -16,51 +16,42 @@
 
 package rpc
 
-import (
-	"compress/gzip"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"strings"
-	"sync"
-)
-
-var gzPool = sync.Pool{
-	New: func() interface{} {
-		w := gzip.NewWriter(ioutil.Discard)
-		return w
-	},
-}
-
-type gzipResponseWriter struct {
-	io.Writer
-	http.ResponseWriter
-}
-
-func (w *gzipResponseWriter) WriteHeader(status int) {
-	w.Header().Del("Content-Length")
-	w.ResponseWriter.WriteHeader(status)
-}
-
-func (w *gzipResponseWriter) Write(b []byte) (int, error) {
-	return w.Writer.Write(b)
-}
-
-func newGzipHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if !strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
-			next.ServeHTTP(w, r)
-			return
-		}
-
-		w.Header().Set("Content-Encoding", "gzip")
-
-		gz := gzPool.Get().(*gzip.Writer)
-		defer gzPool.Put(gz)
-
-		gz.Reset(w)
-		defer gz.Close()
-
-		next.ServeHTTP(&gzipResponseWriter{ResponseWriter: w, Writer: gz}, r)
-	})
-}
+//var gzPool = sync.Pool{
+//	New: func() interface{} {
+//		w := gzip.NewWriter(ioutil.Discard)
+//		return w
+//	},
+//}
+//
+//type gzipResponseWriter struct {
+//	io.Writer
+//	http.ResponseWriter
+//}
+//
+//func (w *gzipResponseWriter) WriteHeader(status int) {
+//	w.Header().Del("Content-Length")
+//	w.ResponseWriter.WriteHeader(status)
+//}
+//
+//func (w *gzipResponseWriter) Write(b []byte) (int, error) {
+//	return w.Writer.Write(b)
+//}
+//
+//func newGzipHandler(next http.Handler) http.Handler {
+//	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+//		if !strings.Contains(r.Header.Get("Accept-Encoding"), "gzip") {
+//			next.ServeHTTP(w, r)
+//			return
+//		}
+//
+//		w.Header().Set("Content-Encoding", "gzip")
+//
+//		gz := gzPool.Get().(*gzip.Writer)
+//		defer gzPool.Put(gz)
+//
+//		gz.Reset(w)
+//		defer gz.Close()
+//
+//		next.ServeHTTP(&gzipResponseWriter{ResponseWriter: w, Writer: gz}, r)
+//	})
+//}
